+++ src/oic/oic/message.py
@@ -38,6 +38,7 @@ from oic.oauth2.message import NotAllowedValue
 from oic.oauth2.message import ParamDefinition
 from oic.oauth2.message import SchemeError
 from oic.utils import time_util
+from oic.utils.time_util import utc_time_sans_frac
 
 __author__ = "rohe0002"
 
@@ -269,10 +270,13 @@ def check_char_set(string, allowed):
             raise NotAllowedValue("'%c' not in the allowed character set" % c)
 
 
+TOKEN_VERIFY_ARGS = ["key", "keyjar", "algs", "sender"]
+
+
 def verify_id_token(instance, check_hash=False, **kwargs):
     # Try to decode the JWT, checks the signature
     args = {}
-    for arg in ["key", "keyjar", "algs", "sender"]:
+    for arg in TOKEN_VERIFY_ARGS:
         try:
             args[arg] = kwargs[arg]
         except KeyError:
@@ -342,8 +346,9 @@ class AccessTokenResponse(message.AccessTokenResponse):
     c_param.update({"id_token": SINGLE_OPTIONAL_STRING})
 
     def verify(self, **kwargs):
-        super(AccessTokenResponse, self).verify(**kwargs)
+        super().verify(**kwargs)
         if "id_token" in self:
+            self.raw_id_token = self["id_token"]
             # replace the JWT with the verified IdToken instance
             self["id_token"] = verify_id_token(self, **kwargs)
 
@@ -367,7 +372,7 @@ class AuthorizationResponse(message.AuthorizationResponse, message.AccessTokenRe
     )
 
     def verify(self, **kwargs):
-        super(AuthorizationResponse, self).verify(**kwargs)
+        super().verify(**kwargs)
 
         if "aud" in self:
             if "client_id" in kwargs:
@@ -376,6 +381,7 @@ class AuthorizationResponse(message.AuthorizationResponse, message.AccessTokenRe
                     return False
 
         if "id_token" in self:
+            self.raw_id_token = self["id_token"]
             self["id_token"] = verify_id_token(self, check_hash=True, **kwargs)
 
         if "access_token" in self:
@@ -444,7 +450,7 @@ class AuthorizationRequest(message.AuthorizationRequest):
         All parameter values that are present both in the OAuth 2.0
         Authorization Request and in the OpenID Request Object MUST exactly match.
         """
-        super(AuthorizationRequest, self).verify(**kwargs)
+        super().verify(**kwargs)
 
         args = {}
         for arg in ["key", "keyjar", "opponent_id", "sender"]:
@@ -550,7 +556,7 @@ class OpenIDSchema(Message):
     }
 
     def verify(self, **kwargs):
-        super(OpenIDSchema, self).verify(**kwargs)
+        super().verify(**kwargs)
 
         if "birthdate" in self:
             # Either YYYY-MM-DD or just YYYY or 0000-MM-DD
@@ -604,6 +610,10 @@ class RegistrationRequest(Message):
         "initiate_login_uri": SINGLE_OPTIONAL_STRING,
         "request_uris": OPTIONAL_LIST_OF_STRINGS,
         "post_logout_redirect_uris": OPTIONAL_LIST_OF_STRINGS,
+        "frontchannel_logout_uri": SINGLE_OPTIONAL_STRING,
+        "frontchannel_logout_session_required": OPTIONAL_LOGICAL,
+        "backchannel_logout_uri": SINGLE_OPTIONAL_STRING,
+        "backchannel_logout_session_required": OPTIONAL_LOGICAL,
     }
     c_default = {"application_type": "web", "response_types": ["code"]}
     c_allowed_values = {
@@ -612,7 +622,7 @@ class RegistrationRequest(Message):
     }
 
     def verify(self, **kwargs):
-        super(RegistrationRequest, self).verify(**kwargs)
+        super().verify(**kwargs)
 
         if "initiate_login_uri" in self and not self["initiate_login_uri"].startswith(
             "https:"
@@ -774,36 +784,60 @@ class IdToken(OpenIDSchema):
         return True
 
 
-class RefreshSessionRequest(Message):
-    c_param = {
-        "id_token": SINGLE_REQUIRED_STRING,
-        "redirect_url": SINGLE_REQUIRED_STRING,
-        "state": SINGLE_REQUIRED_STRING,
-    }
+class StateFullMessage(Message):
+    c_param = {"state": SINGLE_REQUIRED_STRING}
 
 
-class RefreshSessionResponse(Message):
-    c_param = {"id_token": SINGLE_REQUIRED_STRING, "state": SINGLE_REQUIRED_STRING}
+class RefreshSessionRequest(StateFullMessage):
+    c_param = StateFullMessage.c_param.copy()
+    c_param.update(
+        {"id_token": SINGLE_REQUIRED_STRING, "redirect_url": SINGLE_REQUIRED_STRING}
+    )
+
+    def verify(self, **kwargs):
+        super(RefreshSessionRequest, self).verify(**kwargs)
+        if "id_token" in self:
+            self.raw_id_token = self["id_token"]
+            self["id_token"] = verify_id_token(self, check_hash=True, **kwargs)
+
+
+class RefreshSessionResponse(StateFullMessage):
+    c_param = StateFullMessage.c_param.copy()
+    c_param.update({"id_token": SINGLE_REQUIRED_STRING})
+
+    def verify(self, **kwargs):
+        super(RefreshSessionResponse, self).verify(**kwargs)
+        if "id_token" in self:
+            self.raw_id_token = self["id_token"]
+            self["id_token"] = verify_id_token(self, check_hash=True, **kwargs)
 
 
 class CheckSessionRequest(Message):
     c_param = {"id_token": SINGLE_REQUIRED_STRING}
+
+    def verify(self, **kwargs):
+        super(CheckSessionRequest, self).verify(**kwargs)
+        if "id_token" in self:
+            self.raw_id_token = self["id_token"]
+            self["id_token"] = verify_id_token(self, check_hash=True, **kwargs)
 
 
 class CheckIDRequest(Message):
     c_param = {"access_token": SINGLE_REQUIRED_STRING}
 
 
-class EndSessionRequest(Message):
-    c_param = {
-        "id_token_hint": SINGLE_OPTIONAL_STRING,
-        "post_logout_redirect_uri": SINGLE_OPTIONAL_STRING,
-        "state": SINGLE_OPTIONAL_STRING,
-    }
+class EndSessionRequest(StateFullMessage):
+    c_param = StateFullMessage.c_param.copy()
+    c_param.update(
+        {
+            "id_token_hint": SINGLE_OPTIONAL_STRING,
+            "post_logout_redirect_uri": SINGLE_OPTIONAL_STRING,
+        }
+    )
 
 
-class EndSessionResponse(Message):
-    c_param = {"state": SINGLE_OPTIONAL_STRING}
+class EndSessionResponse(StateFullMessage):
+    pass
 
 
 class Claims(Message):
@@ -872,7 +906,7 @@ class ProviderConfigurationResponse(Message):
     }
 
     def verify(self, **kwargs):
-        super(ProviderConfigurationResponse, self).verify(**kwargs)
+        super().verify(**kwargs)
 
         if "scopes_supported" in self:
             if "openid" not in self["scopes_supported"]:
@@ -980,6 +1014,112 @@ SCOPE2CLAIMS = {
     "offline_access": [],
 }  # type: Dict[str, List[str]]
 
+# LOGOUT related messages
+
+SINGLE_OPTIONAL_JSON = ParamDefinition(dict, False, json_ser, json_deser, False)
+SINGLE_REQUIRED_JSON = ParamDefinition(dict, True, json_ser, json_deser, False)
+
+BACK_CHANNEL_LOGOUT_EVENT = "http://schemas.openid.net/event/backchannel-logout"
+
+
+class LogoutToken(Message):
+    """Defined in https://openid.net/specs/openid-connect-backchannel-1_0.html#LogoutToken ."""
+
+    c_param = {
+        "iss": SINGLE_REQUIRED_STRING,
+        "sub": SINGLE_OPTIONAL_STRING,
+        "aud": REQUIRED_LIST_OF_STRINGS,  # Array of strings or string
+        "iat": SINGLE_REQUIRED_INT,
+        "jti": SINGLE_REQUIRED_STRING,
+        "events": SINGLE_REQUIRED_JSON,
+        "sid": SINGLE_OPTIONAL_STRING,
+    }
+
+    def verify(self, **kwargs):
+        super().verify(**kwargs)
+
+        if "nonce" in self:
+            raise MessageException(
+                '"nonce" is prohibited from appearing in a LogoutToken.'
+            )
+
+        # Check the 'events' JSON
+        _keys = list(self["events"].keys())
+        if len(_keys) != 1:
+            raise ValueError('Must only be one member in "events"')
+        if _keys[0] != BACK_CHANNEL_LOGOUT_EVENT:
+            raise ValueError('Wrong member in "events"')
+        if self["events"][_keys[0]] != {}:
+            raise ValueError('Wrong member value in "events"')
+
+        # There must be either a 'sub' or a 'sid', and may contain both
+        if not ("sub" in self or "sid" in self):
+            raise ValueError('There MUST be either a "sub" or a "sid"')
+
+        try:
+            if kwargs["aud"] not in self["aud"]:
+                raise NotForMe("Not among intended audience")
+        except KeyError:
+            pass
+
+        try:
+            if kwargs["iss"] != self["iss"]:
+                raise NotForMe("Wrong issuer")
+        except KeyError:
+            pass
+
+        _now = utc_time_sans_frac()
+
+        _skew = kwargs.get("skew", 0)
+        _iat = self.get("iat", 0)
+
+        if _iat and _iat > (_now + _skew):
+            raise ValueError("Invalid issued_at time")
+
+        return True
+
+
+ID_TOKEN_VERIFY_ARGS = [
+    "keyjar",
+    "verify",
+    "encalg",
+    "encenc",
+    "sigalg",
+    "issuer",
+    "allow_missing_kid",
+    "no_kid_issuer",
+    "trusting",
+    "skew",
+    "nonce_storage_time",
+    "client_id",
+]
+
+
+class BackChannelLogoutRequest(Message):
+    """Defines the message used in https://openid.net/specs/openid-connect-backchannel-1_0.html ."""
+
+    c_param = {"logout_token": SINGLE_REQUIRED_STRING}
+
+    def verify(self, **kwargs):
+        super().verify(**kwargs)
+
+        args = {arg: kwargs[arg] for arg in TOKEN_VERIFY_ARGS if arg in kwargs}
+
+        logout_token = LogoutToken().from_jwt(str(self["logout_token"]), **args)
+        logout_token.verify(**kwargs)
+
+        self["logout_token"] = logout_token
+        logger.info("Verified Logout Token: {}".format(logout_token.to_dict()))
+
+        return True
+
+
+class FrontChannelLogoutRequest(Message):
+    """Defines the message used in https://openid.net/specs/openid-connect-frontchannel-1_0.html ."""
+
+    c_param = {"iss": SINGLE_OPTIONAL_STRING, "sid": SINGLE_OPTIONAL_STRING}
+
+
 MSG = {
     "RefreshAccessTokenRequest": RefreshAccessTokenRequest,
     "TokenErrorResponse": TokenErrorResponse,
@@ -1009,6 +1149,10 @@ MSG = {
     "DiscoveryRequest": DiscoveryRequest,
     "DiscoveryResponse": DiscoveryResponse,
     "ResourceRequest": ResourceRequest,
+    # LOGOUT messages
+    "LogoutToken": LogoutToken,
+    "BackChannelLogoutRequest": BackChannelLogoutRequest,
+    "FrontChannelLogoutRequest": FrontChannelLogoutRequest,
 }
 
 
@@ -1042,7 +1186,6 @@ class OIDCMessageFactory(MessageFactory):
     endsession_endpoint = MessageTuple(EndSessionRequest, EndSessionResponse)
     checkid_endpoint = MessageTuple(CheckIDRequest, IdToken)
     checksession_endpoint = MessageTuple(CheckSessionRequest, IdToken)
-    endsession_endpoint = MessageTuple(EndSessionRequest, EndSessionResponse)
     refreshsession_endpoint = MessageTuple(
         RefreshSessionRequest, RefreshSessionResponse
     )